pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wertigo"
version = "0.1.0"
description = "WerTigo Travel Planning and Ticket Tracking System"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "WerTigo Team"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "flask>=2.0.1",
    "flask-cors>=3.0.10",
    "pandas>=1.3.0",
    "numpy>=1.20.0",
    "torch>=1.10.0",
    "transformers>=4.15.0",
    "scikit-learn>=0.24.0",
    "nltk>=3.6.0",
    "tqdm>=4.60.0",
    "matplotlib>=3.4.0",
    "requests>=2.28.0",
    "mysql-connector-python>=8.0.28",
    "spacy>=3.1.0",
    "textblob>=0.15.3",
    "vaderSentiment>=3.3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
]

[tool.setuptools]
package-dir = {"" = "utils"}

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 100 